Harden desktop bridge and layout diff state
