+++ src/app/app-shell/app-shell-layout.tsx
@@ -1,11 +1,10 @@
 import { useCallback, useEffect, useLayoutEffect, useRef, useState } from 'react'
 import { getPersistedSessionPath, isLocalSessionPath } from '../../../shared/session-paths'
-import { defaultDiffBaseline } from '../components/workspace/composer/diff-baseline'
-import type { ProjectDiffBaseline, ProjectDiffRenderMode } from '../desktop/types'
 import { useAnimatedPresence } from '../hooks/useAnimatedPresence'
 import { AppShellLayoutView } from './app-shell-layout-view'
 import { useAppKeybindings } from './useAppKeybindings'
 import type { AppShellController } from './useAppShellController'
+import { useAppShellDiffPreferences } from './useAppShellDiffPreferences'
 import { useAppShellLayoutState } from './useAppShellLayoutState'
 
 type TakeoverTerminalKeyState = {
@@ -30,52 +29,6 @@ function isLocalToPersistedTakeoverTransition(
   )
 }
 
-function areDiffBaselinesEqual(left: ProjectDiffBaseline, right: ProjectDiffBaseline) {
-  if (left.kind !== right.kind) {
-    return false
-  }
-
-  if (left.kind === 'commit' && right.kind === 'commit') {
-    return left.sha === right.sha
-  }
-
-  if (left.kind === 'last-opened' && right.kind === 'last-opened') {
-    return left.rev === right.rev
-  }
-
-  return true
-}
-
-function isSameDraftPromotion({
-  activeThreadId,
-  messageCount,
-  previousSessionPath,
-  previousThreadId,
-  nextSessionPath,
-}: {
-  activeThreadId: string | null
-  messageCount: number | null
-  previousSessionPath: string | null
-  previousThreadId: string | null
-  nextSessionPath: string | null
-}) {
-  return (
-    isLocalSessionPath(previousSessionPath) &&
-    previousThreadId?.startsWith('local-thread-') &&
-    activeThreadId !== null &&
-    getPersistedSessionPath(nextSessionPath) !== null &&
-    (messageCount === null || messageCount <= 1)
-  )
-}
-
-type DiffBaselineState = {
-  projectId: string
-  threadId: string | null
-  sessionPath: string | null
-  baseline: ProjectDiffBaseline
-  source: 'init' | 'override' | 'default'
-}
-
 function getThreadSessionPath(state: AppShellController['state']) {
   if (state.activeView === 'chat' || state.activeView === 'thread' || state.activeView === 'gitops')
     return state.selectedSessionPath
@@ -95,146 +48,6 @@ function isUtilityView(activeView: AppShellController['state']['activeView']) {
     activeView === 'skills' ||
     activeView === 'archived'
   )
-}
-
-function getNextDiffBaseline(controller: AppShellController) {
-  return (
-    controller.activeThreadData?.diffPreferences?.baseline ??
-    controller.shellState?.appSettings.gitDiffBaselineDefault ??
-    defaultDiffBaseline
-  )
-}
-
-function promoteDiffBaselineDraft(options: {
-  activeThreadId: string | null
-  composerProjectId: string
-  controllerRef: React.RefObject<AppShellController>
-  current: DiffBaselineState
-  terminalSessionPath: string | null
-}) {
-  const appDefault = options.controllerRef.current.shellState?.appSettings.gitDiffBaselineDefault
-  const promotedBaseline =
-    appDefault && areDiffBaselinesEqual(options.current.baseline, appDefault)
-      ? null
-      : options.current.baseline
-  void options.controllerRef.current.handleAction('workspace.diff-preferences', {
-    diffBaseline: promotedBaseline,
-  })
-  return {
-    ...options.current,
-    threadId: options.activeThreadId,
-    sessionPath: options.terminalSessionPath,
-  }
-}
-
-function nextDiffBaselineState(options: {
-  activeThreadId: string | null
-  composerProjectId: string
-  controller: AppShellController
-  controllerRef: React.RefObject<AppShellController>
-  current: DiffBaselineState
-  terminalSessionPath: string | null
-}) {
-  const nextBaseline = getNextDiffBaseline(options.controller)
-  if (
-    options.current.projectId === options.composerProjectId &&
-    options.current.source === 'override' &&
-    isSameDraftPromotion({
-      activeThreadId: options.activeThreadId,
-      messageCount: options.controller.activeThreadData?.messages.length ?? null,
-      previousSessionPath: options.current.sessionPath,
-      previousThreadId: options.current.threadId,
-      nextSessionPath: options.terminalSessionPath,
-    })
-  )
-    return promoteDiffBaselineDraft(options)
-  if (
-    options.current.projectId === options.composerProjectId &&
-    options.current.threadId === options.activeThreadId &&
-    options.current.sessionPath === options.terminalSessionPath &&
-    (options.current.source === 'override' ||
-      areDiffBaselinesEqual(options.current.baseline, nextBaseline))
-  )
-    return options.current
-  return {
-    projectId: options.composerProjectId,
-    threadId: options.activeThreadId,
-    sessionPath: options.terminalSessionPath,
-    baseline: nextBaseline,
-    source: 'init' as const,
-  }
-}
-
-type DiffRenderModeState = {
-  projectId: string
-  threadId: string | null
-  sessionPath: string | null
-  renderMode: ProjectDiffRenderMode
-  source: 'init' | 'override' | 'default'
-}
-
-function getNextDiffRenderMode(controller: AppShellController) {
-  return (
-    controller.activeThreadData?.diffPreferences?.renderMode ??
-    controller.shellState?.appSettings.gitDiffRenderModeDefault ??
-    'stacked'
-  )
-}
-
-function promoteDiffRenderModeDraft(options: {
-  activeThreadId: string | null
-  controllerRef: React.RefObject<AppShellController>
-  current: DiffRenderModeState
-  terminalSessionPath: string | null
-}) {
-  const appDefault = options.controllerRef.current.shellState?.appSettings.gitDiffRenderModeDefault
-  const promotedRenderMode =
-    appDefault === options.current.renderMode ? null : options.current.renderMode
-  void options.controllerRef.current.handleAction('workspace.diff-preferences', {
-    diffRenderMode: promotedRenderMode,
-  })
-  return {
-    ...options.current,
-    threadId: options.activeThreadId,
-    sessionPath: options.terminalSessionPath,
-  }
-}
-
-function nextDiffRenderModeState(options: {
-  activeThreadId: string | null
-  composerProjectId: string
-  controller: AppShellController
-  controllerRef: React.RefObject<AppShellController>
-  current: DiffRenderModeState
-  terminalSessionPath: string | null
-}) {
-  const nextRenderMode = getNextDiffRenderMode(options.controller)
-  if (
-    options.current.projectId === options.composerProjectId &&
-    options.current.source === 'override' &&
-    isSameDraftPromotion({
-      activeThreadId: options.activeThreadId,
-      messageCount: options.controller.activeThreadData?.messages.length ?? null,
-      previousSessionPath: options.current.sessionPath,
-      previousThreadId: options.current.threadId,
-      nextSessionPath: options.terminalSessionPath,
-    })
-  )
-    return promoteDiffRenderModeDraft(options)
-  if (
-    options.current.projectId === options.composerProjectId &&
-    options.current.threadId === options.activeThreadId &&
-    options.current.sessionPath === options.terminalSessionPath &&
-    (options.current.source === 'override' || options.current.renderMode === nextRenderMode)
-  )
-    return options.current
-  return {
-    projectId: options.composerProjectId,
-    threadId: options.activeThreadId,
-    sessionPath: options.terminalSessionPath,
-    renderMode: nextRenderMode,
-    source: 'init' as const,
-  }
 }
 
 function updateTakeoverTerminalKey(options: {
@@ -266,38 +79,6 @@ function updateTakeoverTerminalKey(options: {
     options.takeoverTerminalKeyRef.current = null
 }
 
-function getEffectiveDiffBaseline(options: {
-  activeThreadId: string | null
-  composerProjectId: string
-  controller: AppShellController
-  diffBaselineState: DiffBaselineState
-  terminalSessionPath: string | null
-}) {
-  if (
-    options.diffBaselineState.projectId === options.composerProjectId &&
-    options.diffBaselineState.threadId === options.activeThreadId &&
-    options.diffBaselineState.sessionPath === options.terminalSessionPath
-  )
-    return options.diffBaselineState.baseline
-  return getNextDiffBaseline(options.controller)
-}
-
-function getEffectiveDiffRenderMode(options: {
-  activeThreadId: string | null
-  composerProjectId: string
-  controller: AppShellController
-  diffRenderModeState: DiffRenderModeState
-  terminalSessionPath: string | null
-}) {
-  if (
-    options.diffRenderModeState.projectId === options.composerProjectId &&
-    options.diffRenderModeState.threadId === options.activeThreadId &&
-    options.diffRenderModeState.sessionPath === options.terminalSessionPath
-  )
-    return options.diffRenderModeState.renderMode
-  return getNextDiffRenderMode(options.controller)
-}
-
 type AppShellLayoutProps = {
   controller: AppShellController
 }
@@ -314,20 +95,6 @@ export function AppShellLayout({ controller }: AppShellLayoutProps) {
   const [closeArtifactDrawerOverlay, setCloseArtifactDrawerOverlay] = useState<(() => void) | null>(
     null,
   )
-  const [diffBaselineState, setDiffBaselineState] = useState<DiffBaselineState>({
-    projectId: '',
-    threadId: null,
-    sessionPath: null,
-    baseline: defaultDiffBaseline,
-    source: 'init',
-  })
-  const [diffRenderModeState, setDiffRenderModeState] = useState<DiffRenderModeState>({
-    projectId: '',
-    threadId: null,
-    sessionPath: null,
-    renderMode: 'stacked',
-    source: 'init',
-  })
   const {
     activeComposerState,
     activeThreadData,
@@ -367,20 +134,13 @@ export function AppShellLayout({ controller }: AppShellLayoutProps) {
     sidebarCompactMode && terminalHiddenByCompactResize
       ? terminalDrawerVisible
       : animatedTerminalDrawerPresent
-  const diffBaseline = getEffectiveDiffBaseline({
-    activeThreadId,
-    composerProjectId,
-    controller,
-    diffBaselineState,
-    terminalSessionPath,
-  })
-  const diffRenderMode = getEffectiveDiffRenderMode({
-    activeThreadId,
-    composerProjectId,
-    controller,
-    diffRenderModeState,
-    terminalSessionPath,
-  })
+  const { diffBaseline, diffRenderMode, handleSetDiffBaseline, handleSetDiffRenderMode } =
+    useAppShellDiffPreferences({
+      activeThreadId,
+      composerProjectId,
+      controller,
+      terminalSessionPath,
+    })
   const { mainSectionRef, takeoverPresent, workspaceContentClass } = useAppShellLayoutState({
     takeoverVisible,
   })
@@ -408,69 +168,6 @@ export function AppShellLayout({ controller }: AppShellLayoutProps) {
 
   const takeoverTerminalKey = takeoverTerminalKeyRef.current?.key ?? nextTakeoverTerminalKey
   controllerRef.current = controller
-
-  useEffect(() => {
-    setDiffBaselineState((current) =>
-      nextDiffBaselineState({
-        activeThreadId,
-        composerProjectId,
-        controller,
-        controllerRef,
-        current,
-        terminalSessionPath,
-      }),
-    )
-  }, [activeThreadId, composerProjectId, controller, terminalSessionPath])
-
-  useEffect(() => {
-    setDiffRenderModeState((current) =>
-      nextDiffRenderModeState({
-        activeThreadId,
-        composerProjectId,
-        controller,
-        controllerRef,
-        current,
-        terminalSessionPath,
-      }),
-    )
-  }, [activeThreadId, composerProjectId, controller, terminalSessionPath])
-
-  const handleSetDiffBaseline = useCallback(
-    (baseline: ProjectDiffBaseline) => {
-      const appDefault = controllerRef.current.shellState?.appSettings.gitDiffBaselineDefault
-      const nextBaseline =
-        appDefault && areDiffBaselinesEqual(baseline, appDefault) ? null : baseline
-      setDiffBaselineState({
-        projectId: composerProjectId,
-        threadId: activeThreadId,
-        sessionPath: terminalSessionPath,
-        baseline,
-        source: nextBaseline ? 'override' : 'default',
-      })
-      void controllerRef.current.handleAction('workspace.diff-preferences', {
-        diffBaseline: nextBaseline,
-      })
-    },
-    [activeThreadId, composerProjectId, terminalSessionPath],
-  )
-
-  const handleSetDiffRenderMode = useCallback(
-    (renderMode: ProjectDiffRenderMode) => {
-      const appDefault = controllerRef.current.shellState?.appSettings.gitDiffRenderModeDefault
-      const nextRenderMode = appDefault === renderMode ? null : renderMode
-      setDiffRenderModeState({
-        projectId: composerProjectId,
-        threadId: activeThreadId,
-        sessionPath: terminalSessionPath,
-        renderMode,
-        source: nextRenderMode ? 'override' : 'default',
-      })
-      void controllerRef.current.handleAction('workspace.diff-preferences', {
-        diffRenderMode: nextRenderMode,
-      })
-    },
-    [activeThreadId, composerProjectId, terminalSessionPath],
-  )
 
   const handleOpenGitOpsFromTakeover = useCallback(async () => {
     controllerRef.current.handleOpenGitOpsView()
