+++ src/app/app-shell/controller-action-utils.ts
@@ -1,6 +1,7 @@
 import { createLocalThreadDraft } from '../../../shared/session-paths'
 import type { DesktopAction } from '../desktop/actions'
 import type { AnyDesktopActionPayload, DesktopActionResult, Thread } from '../desktop/types'
+import { hasDesktopBridgeQuery } from '../query/desktop-query'
 
 export type ActionPayload = AnyDesktopActionPayload
 
@@ -69,11 +70,7 @@ export function sortPinnedProjects<T extends { id: string; pinned?: boolean | un
 }
 
 export function hasDesktopBridge() {
-  if (typeof window === 'undefined') {
-    return false
-  }
-
-  return typeof window.piDesktop?.invokeAction === 'function'
+  return hasDesktopBridgeQuery()
 }
 
 export function buildLocalThreadFallback(
