+++ src/app/hooks/useDesktopTerminal.ts
@@ -3,43 +3,46 @@ import type {
   TerminalEvent,
   TerminalOpenRequest,
   TerminalResizeRequest,
-  TerminalSessionFileStat,
-  TerminalSessionSnapshot,
 } from '../desktop/types'
+import {
+  closeDesktopTerminalQuery,
+  getDesktopTerminalStatusQuery,
+  listDesktopTerminalsQuery,
+  openDesktopTerminalQuery,
+  resizeDesktopTerminalQuery,
+  statDesktopTerminalSessionFileQuery,
+  subscribeDesktopTerminalQuery,
+  writeDesktopTerminalQuery,
+} from '../query/desktop-query'
 
 export async function listDesktopTerminals() {
-  return (await window.piDesktop?.listTerminals?.()) ?? []
+  return listDesktopTerminalsQuery()
 }
 
 export async function openDesktopTerminal(request: TerminalOpenRequest) {
-  if (!window.piDesktop?.openTerminal) {
-    return null as TerminalSessionSnapshot | null
-  }
-
-  return window.piDesktop.openTerminal(request)
+  return openDesktopTerminalQuery(request)
 }
 
 export async function writeDesktopTerminal(sessionId: string, data: string) {
-  await window.piDesktop?.writeTerminal?.(sessionId, data)
+  await writeDesktopTerminalQuery(sessionId, data)
 }
 
 export async function resizeDesktopTerminal(request: TerminalResizeRequest) {
-  await window.piDesktop?.resizeTerminal?.(request)
+  await resizeDesktopTerminalQuery(request)
 }
 
 export async function closeDesktopTerminal(request: TerminalCloseRequest) {
-  await window.piDesktop?.closeTerminal?.(request)
+  await closeDesktopTerminalQuery(request)
 }
 
 export async function statDesktopTerminalSessionFile(sessionId: string) {
-  return ((await window.piDesktop?.statTerminalSessionFile?.(sessionId)) ??
-    null) as TerminalSessionFileStat | null
+  return statDesktopTerminalSessionFileQuery(sessionId)
 }
 
 export async function getDesktopTerminalStatus(sessionId: string) {
-  return (await window.piDesktop?.getTerminalStatus?.(sessionId)) ?? null
+  return getDesktopTerminalStatusQuery(sessionId)
 }
 
 export function subscribeDesktopTerminal(listener: (event: TerminalEvent) => void) {
-  return window.piDesktop?.subscribeTerminal?.(listener) ?? (() => undefined)
+  return subscribeDesktopTerminalQuery(listener)
 }
